challenge-django-consultant/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
ipdb==0.13.13
requests==2.22.0
celery==5.3.1
redis==5.0.0